+++ Sources/VComponents/Components/Containers/Disclosure Group/VDisclosureGroup.swift
@@ -22,7 +22,7 @@ import VCore
 ///             content: {
 ///                 LazyVStack(spacing: 0, content: {
 ///                     ForEach(0..<10, content: { num in
-///                         VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+///                         VListRow(uiModel: .noFirstAndLastSeparators(isFirst: num == 0), content: {
 ///                             Text(String(num))
 ///                                 .frame(maxWidth: .infinity, alignment: .leading)
 ///                         })
@@ -225,7 +225,7 @@ struct VDisclosureGroup_Previews: PreviewProvider {
                     content: {
                         LazyVStack(spacing: 0, content: {
                             ForEach(0..<10, content: { num in
-                                VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+                                VListRow(uiModel: .noFirstAndLastSeparators(isFirst: num == 0), content: {
                                     Text(String(num))
                                         .frame(maxWidth: .infinity, alignment: .leading)
                                 })