+++ Sources/VComponents/Components/Modals/Bottom Sheet/VBottomSheetExtension.swift
@@ -31,7 +31,7 @@ extension View {
     ///                 content: {
     ///                     List(content: {
     ///                         ForEach(0..<20, content: { num in
-    ///                             VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+    ///                             VListRow(uiModel: .noFirstAndLastSeparators(isFirst: num == 0), content: {
     ///                                 Text(String(num))
     ///                                     .frame(maxWidth: .infinity, alignment: .leading)
     ///                             })
@@ -95,7 +95,7 @@ extension View {
     ///                 content: { item in
     ///                     List(content: {
     ///                         ForEach(0..<20, content: { num in
-    ///                             VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+    ///                             VListRow(uiModel: .noFirstAndLastSeparators(isFirst: num == 0), content: {
     ///                                 Text(String(num))
     ///                                     .frame(maxWidth: .infinity, alignment: .leading)
     ///                             })