+++ Sources/VComponents/Components/Modals/Modal/VModal.swift
@@ -210,7 +210,7 @@ struct VModal_Previews: PreviewProvider {
             content: {
                 List(content: {
                     ForEach(0..<20, content: { num in
-                        VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+                        VListRow(uiModel: .noFirstAndLastSeparators(isFirst: num == 0), content: {
                             Text(String(num))
                                 .frame(maxWidth: .infinity, alignment: .leading)
                         })