+++ Sources/VComponents/Components/Modals/SideBar/VSideBar.swift
@@ -220,7 +220,7 @@ struct VSideBar_Previews: PreviewProvider {
             content: {
                 List(content: {
                     ForEach(0..<20, content: { num in
-                        VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+                        VListRow(uiModel: .noFirstAndLastSeparators(isFirst: num == 0), content: {
                             Text(String(num))
                                 .frame(maxWidth: .infinity, alignment: .leading)
                         })