+++ Sources/VComponents/Components/Modals/SideBar/VSideBarExtension.swift
@@ -30,7 +30,7 @@ extension View {
     ///                 content: {
     ///                     List(content: {
     ///                         ForEach(0..<20, content: { num in
-    ///                             VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+    ///                             VListRow(uiModel: .noFirstAndLastSeparators(isFirst: num == 0), content: {
     ///                                 Text(String(num))
     ///                                     .frame(maxWidth: .infinity, alignment: .leading)
     ///                             })
@@ -92,7 +92,7 @@ extension View {
     ///                 content: { item in
     ///                     List(content: {
     ///                         ForEach(0..<20, content: { num in
-    ///                             VListRow(separator: .noFirstAndLastSeparators(isFirst: num == 0), content: {
+    ///                             VListRow(uiModel: .noFirstAndLastSeparators(isFirst: num == 0), content: {
     ///                                 Text(String(num))
     ///                                     .frame(maxWidth: .infinity, alignment: .leading)
     ///                             })